requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
requests==2.31.0
schedule==1.2.1
speedtest-cli==2.1.3
urllib3==2.2.1